long::open
