+++ bindings/hyperdrivepy/src/hyperdrive_state_methods.rs
@@ -26,33 +26,6 @@ impl HyperdriveState {
         Ok(result)
     }
 
-    pub fn calculate_spot_price_after_long(
-        &self,
-        base_amount: &str,
-        maybe_bond_amount: Option<&str>,
-    ) -> PyResult<String> {
-        let base_amount_fp = FixedPoint::from(U256::from_dec_str(base_amount).map_err(|_| {
-            PyErr::new::<PyValueError, _>("Failed to convert base_amount string to U256")
-        })?);
-        let maybe_bond_amount_fp = if let Some(bond_amount) = maybe_bond_amount {
-            Some(FixedPoint::from(U256::from_dec_str(bond_amount).map_err(
-                |_| {
-                    PyErr::new::<PyValueError, _>(
-                        "Failed to convert maybe_bond_amount string to U256",
-                    )
-                },
-            )?))
-        } else {
-            None
-        };
-        let result_fp = self
-            .state
-            .calculate_spot_price_after_long(base_amount_fp, maybe_bond_amount_fp)
-            .unwrap();
-        let result = U256::from(result_fp).to_string();
-        Ok(result)
-    }
-
     pub fn calculate_spot_price_after_short(
         &self,
         bond_amount: &str,
@@ -87,33 +60,6 @@ impl HyperdriveState {
         Ok(result)
     }
 
-    pub fn calculate_spot_rate_after_long(
-        &self,
-        base_amount: &str,
-        maybe_bond_amount: Option<&str>,
-    ) -> PyResult<String> {
-        let base_amount_fp = FixedPoint::from(U256::from_dec_str(base_amount).map_err(|_| {
-            PyErr::new::<PyValueError, _>("Failed to convert base_amount string to U256")
-        })?);
-        let maybe_bond_amount_fp = if let Some(bond_amount) = maybe_bond_amount {
-            Some(FixedPoint::from(U256::from_dec_str(bond_amount).map_err(
-                |_| {
-                    PyErr::new::<PyValueError, _>(
-                        "Failed to convert maybe_bond_amount string to U256",
-                    )
-                },
-            )?))
-        } else {
-            None
-        };
-        let result_fp = self
-            .state
-            .calculate_spot_rate_after_long(base_amount_fp, maybe_bond_amount_fp)
-            .unwrap();
-        let result = U256::from(result_fp).to_string();
-        Ok(result)
-    }
-
     pub fn calculate_spot_rate(&self) -> PyResult<String> {
         let result_fp = self.state.calculate_spot_rate();
         let result = U256::from(result_fp).to_string();
@@ -140,32 +86,6 @@ impl HyperdriveState {
         let result2 = result_fp2.to_string();
         let result3 = U256::from(result_fp3).to_string();
         Ok((result1, result2, result3))
-    }
-
-    pub fn calculate_open_long(&self, base_amount: &str) -> PyResult<String> {
-        let base_amount_fp = FixedPoint::from(U256::from_dec_str(base_amount).map_err(|_| {
-            PyErr::new::<PyValueError, _>("Failed to convert base_amount string to U256")
-        })?);
-        let result_fp = self.state.calculate_open_long(base_amount_fp).unwrap();
-        let result = U256::from(result_fp).to_string();
-        Ok(result)
-    }
-
-    pub fn calculate_pool_deltas_after_open_long(&self, base_amount: &str) -> PyResult<String> {
-        let base_amount_fp = FixedPoint::from(U256::from_dec_str(base_amount).map_err(|_| {
-            PyErr::new::<PyValueError, _>("Failed to convert base_amount string to U256")
-        })?);
-        let result_fp = self
-            .state
-            .calculate_pool_deltas_after_open_long(base_amount_fp)
-            .map_err(|err| {
-                PyErr::new::<PyValueError, _>(format!(
-                    "calculate_pool_deltas_after_open_long returned the error: {:?}",
-                    err
-                ))
-            })?;
-        let result = U256::from(result_fp).to_string();
-        Ok(result)
     }
 
     pub fn calculate_open_short(
